requirements.txt
Flask==2.1.0
gunicorn==20.1.0
flask_ngrok==0.0.25
werkzeug==2.1.2
numpy==1.22.4
tensorflow==2.9.0
Pillow==9.1.1
protobuf~=3.19.0
google-cloud-storage